fix duplicated code
